requirements.txt
altair==4.1.0
click==7.1.2
gtts==2.2.3
python-dotenv==0.19.0
SpeechRecognition==3.8.1
sounddevice==0.4.1
soundfile==0.11.0
streamlit==1.7.0
vega_datasets==0.9.0
google-api-core==2.19.0
googleapis-common-protos==1.63.0
grpcio-status==1.62.2
protobuf==4.21.6
google-generativeai==0.6.4
